refactor(generate pairings and update score)
- Change the roundPairings field on the Tournament model to be a map
of objects. It needs to be this in order for us to have dynamic key
values.

- Update the logic for generating pairings and updating the score. Because
of the switch of roundPairings to be a map, the data has to be accessed
in a new way. Values to be read have to be accessed with the .get() method,
and values to be written need to be done so with the .set() method.